Create working timer start endpoint and logic
