+++ index.js
@@ -5,4 +5,4 @@ const port = process.env.PORT || 8000;
 
 app.listen(port, () => console.log(`\n** server is on port ${port} **\n`));
 
-focusbot.createBot();
+//focusbot.createBot();
